Split value store related types to separate files
