+++ toolchain/check/BUILD
@@ -119,6 +119,7 @@ cc_library(
         "//common:raw_string_ostream",
         "//common:set",
         "//common:vlog",
+        "//toolchain/base:canonical_value_store",
         "//toolchain/base:index_base",
         "//toolchain/base:kind_switch",
         "//toolchain/base:value_store",
@@ -267,6 +268,7 @@ cc_library(
         "//common:ostream",
         "//common:set",
         "//common:vlog",
+        "//toolchain/base:canonical_value_store",
         "//toolchain/base:index_base",
         "//toolchain/base:value_ids",
         "//toolchain/base:value_store",
