+++ toolchain/sem_ir/BUILD
@@ -10,8 +10,10 @@ cc_library(
     name = "block_value_store",
     hdrs = ["block_value_store.h"],
     deps = [
+        "//common:check",
         "//common:hashing",
         "//common:set",
+        "//toolchain/base:mem_usage",
         "//toolchain/base:value_store",
         "//toolchain/base:yaml",
         "@llvm-project//llvm:Support",
@@ -131,9 +133,11 @@ cc_library(
         "//common:raw_string_ostream",
         "//common:set",
         "//common:struct_reflection",
+        "//toolchain/base:canonical_value_store",
         "//toolchain/base:index_base",
         "//toolchain/base:int",
         "//toolchain/base:kind_switch",
+        "//toolchain/base:relational_value_store",
         "//toolchain/base:shared_value_stores",
         "//toolchain/base:value_ids",
         "//toolchain/base:value_store",
